refactor: move issuance internals into workspace package
Move the mosaic service (service/types/utils) and issuance templates into
a new private @sdp/issuance workspace package, following the sdp-rpc/
solana/custody/payments extraction pattern. Thin re-export shims remain
at the old apps/sdp-api paths so no consumer imports change. The
createMosaicService factory stays app-side, wiring the fee-payment
adapter and AppError mapping.
