+++ apps/sdp-api/Dockerfile
@@ -23,6 +23,7 @@ COPY packages/sdp-payments/package.json ./packages/sdp-payments/
 COPY packages/sdp-rpc/package.json ./packages/sdp-rpc/
 COPY packages/sdp-custody/package.json ./packages/sdp-custody/
 COPY packages/sdp-solana/package.json ./packages/sdp-solana/
+COPY packages/sdp-issuance/package.json ./packages/sdp-issuance/
 COPY patches ./patches
 
 RUN --mount=type=cache,id=pnpm-store,target=/pnpm/store \
@@ -39,6 +40,7 @@ COPY packages/sdp-payments ./packages/sdp-payments
 COPY packages/sdp-rpc ./packages/sdp-rpc
 COPY packages/sdp-custody ./packages/sdp-custody
 COPY packages/sdp-solana ./packages/sdp-solana
+COPY packages/sdp-issuance ./packages/sdp-issuance
 
 WORKDIR /repo/apps/sdp-api
 RUN node scripts/build-node.mjs
