+++ apps/sdp-api/Dockerfile.dockerignore
@@ -23,6 +23,7 @@ packages/*
 !packages/sdp-custody
 !packages/sdp-solana
 !packages/sdp-payments
+!packages/sdp-issuance
 docs
 infra
 scripts
