+++ apps/sdp-docs/Dockerfile
@@ -23,6 +23,7 @@ COPY packages/sdp-payments/package.json ./packages/sdp-payments/
 COPY packages/sdp-rpc/package.json ./packages/sdp-rpc/
 COPY packages/sdp-custody/package.json ./packages/sdp-custody/
 COPY packages/sdp-solana/package.json ./packages/sdp-solana/
+COPY packages/sdp-issuance/package.json ./packages/sdp-issuance/
 COPY patches ./patches
 
 RUN --mount=type=cache,id=pnpm-store,target=/pnpm/store \
@@ -40,6 +41,7 @@ COPY packages/sdp-payments ./packages/sdp-payments
 COPY packages/sdp-rpc ./packages/sdp-rpc
 COPY packages/sdp-custody ./packages/sdp-custody
 COPY packages/sdp-solana ./packages/sdp-solana
+COPY packages/sdp-issuance ./packages/sdp-issuance
 
 # These are parsed as URLs at build time, so the placeholder defaults are valid
 # URLs whose origin docker/entrypoint.sh swaps for runtime env at container start.
