+++ packages/sdp-issuance/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "@sdp/issuance",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "default": "./src/index.ts"
+    },
+    "./mosaic": {
+      "types": "./src/mosaic/index.ts",
+      "import": "./src/mosaic/index.ts",
+      "default": "./src/mosaic/index.ts"
+    },
+    "./mosaic/service": {
+      "types": "./src/mosaic/service.ts",
+      "import": "./src/mosaic/service.ts",
+      "default": "./src/mosaic/service.ts"
+    },
+    "./mosaic/types": {
+      "types": "./src/mosaic/types.ts",
+      "import": "./src/mosaic/types.ts",
+      "default": "./src/mosaic/types.ts"
+    },
+    "./mosaic/utils": {
+      "types": "./src/mosaic/utils.ts",
+      "import": "./src/mosaic/utils.ts",
+      "default": "./src/mosaic/utils.ts"
+    },
+    "./templates": {
+      "types": "./src/templates/index.ts",
+      "import": "./src/templates/index.ts",
+      "default": "./src/templates/index.ts"
+    },
+    "./templates/definitions": {
+      "types": "./src/templates/definitions.ts",
+      "import": "./src/templates/definitions.ts",
+      "default": "./src/templates/definitions.ts"
+    }
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "lint": "biome check src/"
+  },
+  "dependencies": {
+    "@sdp/payments": "workspace:*",
+    "@sdp/rpc": "workspace:*",
+    "@sdp/solana": "workspace:*",
+    "@sdp/types": "workspace:*",
+    "@solana-program/system": "catalog:",
+    "@solana-program/token-2022": "catalog:",
+    "@solana/kit": "catalog:",
+    "@solana/mosaic-sdk": "catalog:",
+    "@solana/signers": "catalog:",
+    "@solana/token-acl-gate-sdk": "0.2.0"
+  }
+}
